增加共享内存

+++ AI-MasterControl/Element/ElementArc.cs
@@ -8,9 +8,9 @@ namespace AI_MasterControl.Element
     {
         public Int32 rx;
         public Int32 ry;
-        public float Start_angle;
-        public float End_anlge;
-        public ElementArc(Int32 type,Int32 x,Int32 y,ElementColor color,Int32 rx,Int32 ry,float start_angle,float end_angle) : base(type, color, x, y)
+        public double Start_angle;
+        public double End_anlge;
+        public ElementArc(Int32 type,Int32 x,Int32 y,ElementColor color,Int32 rx,Int32 ry,double start_angle,double end_angle) : base(type, color, x, y)
         {
             this.rx = rx;
             this.ry = ry;
